requirements.txt
-i https://pypi.org/simple
contourpy==1.3.2; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
fonttools==4.58.4; python_version >= '3.9'
kiwisolver==1.4.8; python_version >= '3.10'
matplotlib==3.10.3; python_version >= '3.10'
numpy==2.3.1; python_version >= '3.11'
packaging==25.0; python_version >= '3.8'
pillow==11.2.1; python_version >= '3.9'
pyparsing==3.2.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
wordcloud==1.9.4; python_version >= '3.7'
